Separate HTTP from the Representor
