+++ Representor/Transition.swift
@@ -41,40 +41,3 @@ public protocol Transition : Equatable, Hashable {
   var attributes:InputProperties { get }
   var parameters:InputProperties { get }
 }
-
-public struct HTTPTransition : Transition {
-  public typealias Builder = HTTPTransitionBuilder
-
-  public let uri:String
-
-  public let attributes:InputProperties
-  public let parameters:InputProperties
-
-  public init(uri:String, attributes:InputProperties, parameters:InputProperties) {
-    self.uri = uri
-    self.attributes = attributes
-    self.parameters = parameters
-  }
-
-  public init(uri:String, _ block:((builder:Builder) -> ())) {
-    let builder = Builder()
-
-    block(builder: builder)
-
-    self.uri = uri
-    self.attributes = builder.attributes
-    self.parameters = builder.parameters
-  }
-
-  public var hashValue:Int {
-    return uri.hashValue
-  }
-}
-
-public func ==(lhs:HTTPTransition, rhs:HTTPTransition) -> Bool {
-  return (
-    lhs.uri == rhs.uri &&
-    lhs.attributes == rhs.attributes &&
-    lhs.parameters == rhs.parameters
-  )
-}
